demo/requirements.txt
Django>=4.1.0,<4.2.0
djangorestframework>=3.14.0,<3.15.0
django-cors-headers>=3.13.0,<3.14.0
django_api_data_cache